Update music.ts
MMR 1.16.0.6 and above allow for ".seq" and ".aseq" files to be used alongside ".zseq" files for an MMRS file. To maintain compatibility with MMRS files that may use those files instead of just ".zseq" files add them as regex alternatives to find those files in the zipped archive instead of just searching for ".zseq" files.